refactor pathing and new dataset method
