--- graphite/dataset/__init__.py
@@ -1,1 +1,0 @@
-from .dataset_generator import MetricTSPGenerator, GeneralTSPGenerator
